+++ src/main/webapp/entries.js
@@ -52,7 +52,8 @@ module.exports = {
   "remote-api-details":
     "./resources/js/pages/remote-apis/remote-api-details.js",
   "remote-apis": "./resources/js/pages/remote-apis/RemoteApiPage.jsx",
-  "create-metadata-template": "./resources/js/components/metadata-template",
+  "create-metadata-template":
+    "./resources/js/components/metadata-template/details",
   "visualizations-phylogenetics":
     "./resources/js/pages/visualizations/phylogenetics/index.js",
   "project-users": "./resources/js/pages/projects/ProjectMembers.jsx",
