+++ src/main/webapp/resources/js/components/metadata-template/details/TemplateFields.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Alert } from "antd";
-import { IconInfoCircle } from "../icons/Icons";
+import { IconInfoCircle } from "../../icons/Icons";
 
 export function TemplateFields() {
   return (
