+++ src/main/webapp/resources/js/components/metadata-template/details/index.js
