+++ src/main/webapp/resources/js/components/metadata-template/details/template-utilities.test.js
